+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/SessionUtils.kt
@@ -230,6 +230,41 @@ class SessionUtilsClient(
         )
     }
 
+    fun idTokenCallback(
+        provider: String,
+        idToken: String,
+        nonce: String,
+        scope: Collection<String>,
+        success: Success<AuthToken>,
+        failure: Failure<ReachFiveError>,
+        origin: String? = null
+    ) {
+        val redirectUri = sdkConfig.scheme
+        val pkce = PkceAuthCodeFlow.generate(redirectUri)
+
+        loginCallbackHandler.getAuthorizationCode(
+            provider = provider,
+            idToken = idToken,
+            nonce = nonce,
+            pkce = pkce,
+            clientId = sdkConfig.clientId,
+            redirectUri = redirectUri,
+            scope = scope,
+            origin = origin,
+            success = { authCode ->
+                exchangeAuthorizationCode(
+                    authCode,
+                    redirectUri,
+                    pkce.codeVerifier,
+                    success,
+                    failure,
+                    sdkMethod = "idTokenCallback"
+                )
+            },
+            failure = failure
+        )
+    }
+
     override fun loginWithWeb(
         scope: Collection<String>,
         state: String?,