+++ sdk-facebook/src/main/java/co/reachfive/identity/sdk/facebook/FacebookProvider.kt
@@ -9,6 +9,7 @@ import co.reachfive.identity.sdk.core.SessionUtilsClient
 import co.reachfive.identity.sdk.core.models.AuthToken
 import co.reachfive.identity.sdk.core.models.ProviderConfig
 import co.reachfive.identity.sdk.core.models.ReachFiveError
+import co.reachfive.identity.sdk.core.models.SdkConfig
 import co.reachfive.identity.sdk.core.utils.Failure
 import co.reachfive.identity.sdk.core.utils.Success
 import co.reachfive.identity.sdk.facebook.FacebookProvider.Companion.REQUEST_CODE
@@ -27,7 +28,8 @@ class FacebookProvider : ProviderCreator {
     override fun create(
         providerConfig: ProviderConfig,
         sessionUtils: SessionUtilsClient,
-        context: Context
+        context: Context,
+        sdkConfig: SdkConfig,
     ): Provider {
         return ConfiguredFacebookProvider(providerConfig, sessionUtils)
     }